Refactor torch._dynamo.utils helper groups
